Tidy the RX/TX pipeline sample-count and sentinel handling (#28)
- WAV writer: compute the data/RIFF byte counts in 64-bit and clamp them
  to the 32-bit header maximum instead of letting a pass longer than the
  ~4 GiB / ~6.2 h (96 kHz mono) format limit wrap to a tiny size and make
  players truncate playback. Documented the limit at the close site.
- Widen the WAV/IQ sample-count snapshots (and their public getters) to
  int64_t at the source, with the lone consumer in cmd_line.c following.
- Drop frames_in_window, which was only ever incremented and never read.
- Use NAN for the observer/anchor/offset sentinels instead of 0.0/0.0.

The fm_modulate phase-leak the checklist flagged is already fixed: the
modulator moved to src/dsp/fm_mod.c with a caller-owned phase accumulator.

Closes #28

+++ src/ui/cmd_line.c
@@ -473,7 +473,7 @@ static void cmd_dispatch(state_t *state)
                     cmd_set_status(state, "spectrum: a render is already in progress");
                 } else {
                     char wav_path[512];
-                    long n_samples = 0;
+                    int64_t n_samples = 0;
                     int  sample_rate = 0;
                     int  wav_active = 0;
                     rx_session_wav_snapshot(state->rx_session,
@@ -483,8 +483,8 @@ static void cmd_dispatch(state_t *state)
                         cmd_set_status(state, "spectrum: no WAV (recording not started yet)");
                     } else {
                         long want  = (long)(duration_s * (double) sample_rate);
-                        long start = n_samples - want;
-                        if (start < 0) { start = 0; want = n_samples; }
+                        long start = (long)(n_samples - want);
+                        if (start < 0) { start = 0; want = (long) n_samples; }
                         if (want <= 0) {
                             cmd_set_status(state, "spectrum: no samples captured yet");
                         } else {
@@ -521,15 +521,15 @@ static void cmd_dispatch(state_t *state)
                             // falls back to the WAV+ffmpeg path when iq_in
                             // is empty.
                             char iq_path[512] = "";
-                            long iq_pairs = 0;
+                            int64_t iq_pairs = 0;
                             int  iq_rate  = 0;
                             rx_session_iq_snapshot(state->rx_session,
                                                    iq_path, sizeof iq_path,
                                                    &iq_pairs, &iq_rate);
                             if (iq_path[0] && iq_pairs > 0 && iq_rate > 0) {
                                 long want_p  = (long)(duration_s * (double) iq_rate);
-                                long start_p = iq_pairs - want_p;
-                                if (start_p < 0) { start_p = 0; want_p = iq_pairs; }
+                                long start_p = (long)(iq_pairs - want_p);
+                                if (start_p < 0) { start_p = 0; want_p = (long) iq_pairs; }
                                 if (want_p > 0) {
                                     snprintf(state->spec_job.iq_in,
                                              sizeof state->spec_job.iq_in, "%s", iq_path);
